Add duplicate counting in process data script. Add overall counts. Refactor script to take in argumets.

+++ airbyte-tests/bin/create-kube-pod.sh
@@ -1,47 +1,60 @@
 #!/usr/bin/env bash
 
 create_only=0
-
-while getopts 'c:h' opt; do
+delete=1
+total_num_pods=100
+while getopts 'c:kh' opt; do
   case "$opt" in
-    a)
+    c)
       echo "Running in create-only mode.."
       create_only=1
+      total_num_pods=$OPTARG
       ;;
-
+    k)
+      echo "Will not delete completed pods.."
+      delete=0
+      ;;
     ?|h)
-      echo "Usage: $(basename $0) [-c]"
+      echo "Usage: $(basename $0) [-c] [-k]"
       exit 1
       ;;
   esac
 done
 shift "$(($OPTIND -1))"
 
-# watch api and pipe to file
-curl -sN http://127.0.0.1:8001/api/v1/namespaces/default/pods\?watch\=1 | \
-  jq -c 'select(.object.status.phase == "Running") | { "name": .object.metadata.name, "namespace": .object.metadata.namespace, "creationTimestamp": .object.metadata.creationTimestamp, "phase": .object.status.phase, "conditions": [ .object.status.conditions | .[] | select(.status == "True") | {(.type): .lastTransitionTime} ] | add } | select(.conditions.Ready)' >> ./data-gke-1.21.txt &
 
+if [ $create_only -eq 0 ]
+then
+  # watch api and pipe to file
+  echo "Watching for events.."
+  curl -sN http://127.0.0.1:8001/api/v1/namespaces/default/pods\?watch\=1 | \
+    jq -c 'select(.object.status.phase == "Running") | { "name": .object.metadata.name, "namespace": .object.metadata.namespace, "creationTimestamp": .object.metadata.creationTimestamp, "phase": .object.status.phase, "conditions": [ .object.status.conditions | .[] | select(.status == "True") | {(.type): .lastTransitionTime} ] | add } | select(.conditions.Ready)' >> ./data-local-100-completed.txt &
+fi
 
+echo "Creating $total_num_pods test pods.."
 i=1
-while [ $i -ne 300 ]
+while [ $i -ne "$total_num_pods" ]
 do
   echo "creating pod number: $i"
   kubectl create -f ./test-pod.yaml
 
   if [ $create_only -eq 0 ]
   then
-      echo "running in clean up mode.."
-      
-      # wait till pod is in complete
-      while [ "$(kubectl get pod | grep -c 'Compl')" -eq 0 ]
+      # wait till pod is in complete; this works well if there is only one pod
+      while [[ "$(kubectl get pod | grep -c "Running")" -eq 1 ]] || [[ "$(kubectl get pod | grep -c "Pending")" -eq 1 ]] || [[ "$(kubectl get pod | grep -c "ContainerCreating")" -eq 1 ]]
       do
         echo 'wait for pod to finish..'
         sleep 2
       done
 
-      # remove old pod to not mess with results; sleep to let Kube cluster rest
-      kubectl get pod | grep Comp | awk '{print $1}' | xargs kubectl delete pod
-      sleep 1;
+      # remove old pod to not mess with results;
+      if [ $delete -eq 1 ]
+      then
+        kubectl get pod | grep Comp | awk '{print $1}' | xargs kubectl delete pod
+      fi
+
+      # sleep to let Kube cluster rest
+      sleep 2;
 
   fi
   i=$((i+1))
